finish users component

+++ client/src/services/userService.js
@@ -1,5 +1,5 @@
-// src/services/userService.js
-import api from './api';
+import api from './api'
+
 
 export const getAllUsers = async () => {
     const response = await api.get('/User');
@@ -7,7 +7,7 @@ export const getAllUsers = async () => {
 };
 
 export const getUserById = async (id) => {
-    const response = await api.get(`/users/${id}`);
+    const response = await api.get(`/User/${id}`);
     return response.data;
 };
 
@@ -18,13 +18,13 @@ export const createUser = async (user) => {
 };
 
 export const updateUser = async (user) => {
+    console.log(user)
     const response = await api.put('/User', user);
     console.log(response.data);
-    
     return response.data;
 };
 
 export const deleteUser = async (id) => {
-    const response = await api.delete(`/User/${id}`);
+    const response = await api.delete(`/User/${id}`,{headers:{Authorization: `Bearer ${localStorage.getItem('token')}`}});  
     return response.data;
-};
+}
